+++ demos/26-girih/GirihDecagon.js
@@ -1,6 +1,9 @@
 /**
+ * @classdesc The decagon tile from the Girih set.
  *
  * @requires Bounds
+ * @requires GirihTile
+ * @requires Polygon
  * @requires Vertex
  *
  * @author   Ikaros Kappler
@@ -9,36 +12,33 @@
  * @date     2015-03-19 Ikaros Kappler (added toSVG()).
  * @modified 2020-10-30 Refactored to work with PlotBoilerplate.
  * @version  2.0.0-alpha
+ * @file GirihDecacon
+ * @public
  **/
 
+/**
+ * @constructor
+ * @extends GirihTile
+ * @name GirihDecagon
+ * @param {Vertex} position
+ * @param {number} size
+ * @param {number} angle
+ */
+var GirihDecagon = function( position, size, angle ) {
 
-var GirihDecagon = function( position, angle, size ) {
-
-    // console.log("before");
-    GirihTile.call( this, size, position, angle, GirihTile.TYPE_DECAGON );
-    //console.log("after");
+    GirihTile.call( this, position, size, angle, GirihTile.TYPE_DECAGON );
     
-    // Init the actual decahedron shape with the passed size   
-    var pointA = new Vertex(0,0);
-    var pointB = pointA;
-    this._addVertex( pointB );
-
-    var theta = Math.PI/2 * (144.0 / 360.0);
-    for( var i = 1; i <= 9; i++ ) {
-	pointA = pointB; // center of rotation
-	pointB = pointB.clone();
-	pointB.x += size;
-	pointB.rotate( i*theta, pointA );
-	this._addVertex( pointB );
-    }
-
-    // Move to center
-    var bounds = Bounds.computeFromVertices( this.vertices );
-    var move   = new Vertex( size/2.0, 
-			     -bounds.height/2.0
-			   );
-    for( var i = 0; i < this.vertices.length; i++ ) {	
-	this.vertices[i].add( move );		
+    // Init the actual decahedron shape with the passed size:
+    // Divide the full circle into 10 sections (we want to make a regular decagon).
+    var theta = (Math.PI*2) / 10.0;
+    // Compute the 'radius' using pythagoras
+    var radius = Math.sqrt(
+	Math.pow(size/2,2)
+	    +
+	    Math.pow( 1/Math.tan(theta/2) * size/2, 2 )
+    );
+    for( var i = 0; i < 10; i++ ) {
+	this.addVertex( position.clone().addY( -radius ).rotate( theta/2 + i*theta, position ) );
     }
     
     this.imageProperties = {
@@ -54,20 +54,36 @@ var GirihDecagon = function( position, angle, size ) {
 
 
     this._buildInnerPolygons( size );
-    this._buildOuterPolygons();       // Important: call AFTER inner polygons were created!
+    this._buildOuterPolygons( size );       // Important: call AFTER inner polygons were created!
   
 };
 
+
+/**
+ * Build the inner polygons.
+ *
+ * @name _buildInnerPolygons
+ * @memberof GirihDecagon
+ * @private
+ * @param {number} edgeLength
+ */
 GirihDecagon.prototype._buildInnerPolygons = function( edgeLength ) {
-    
+
+    console.log( this.position );
     var centralStar = new Polygon();
     for( var i = 0; i < 10; i++ ) {
 	var innerTile = new Polygon();
 	// Make polygon
 	var topPoint    = this.getVertexAt( i ).clone().scale( 0.5, this.getVertexAt(i+1) );
+	/*
 	var bottomPoint = topPoint.clone().multiplyScalar( 0.615 );
 	var leftPoint   = this.getVertexAt( i ).clone().multiplyScalar( 0.69 );
 	var rightPoint  = this.getVertexAt( i+1 ).clone().multiplyScalar( 0.69 );
+	*/
+
+	var bottomPoint = topPoint.clone().scale( 0.615, this.position );
+	var leftPoint   = this.getVertexAt( i ).clone().scale( 0.69, this.position );
+	var rightPoint  = this.getVertexAt( i+1 ).clone().scale( 0.69, this.position );
 	
 	innerTile.addVertex( topPoint );
 	innerTile.addVertex( rightPoint );
@@ -85,34 +101,32 @@ GirihDecagon.prototype._buildInnerPolygons = function( edgeLength ) {
 };
 
 
+/**
+ * Build the outer polygons.
+ *
+ * @name _buildOuterPolygons
+ * @memberof GirihDecagon
+ * @private
+ * @param {number} edgeLength
+ */
 GirihDecagon.prototype._buildOuterPolygons = function( edgeLength ) {
-
     // DON'T include the inner star here!
     for( var i = 0; i < 10; i++ ) {
-
-	//if( i > 0 )
-	//    continue;
-	
-	//window.alert( this.getInnerTilePolygonAt );
-
 	var outerTile = new Polygon();
 	outerTile.addVertex( this.getVertexAt(i).clone() );
 	outerTile.addVertex( this.innerTilePolygons[i].vertices[0].clone() );
 	outerTile.addVertex( this.innerTilePolygons[i].vertices[3].clone() );
 	outerTile.addVertex( this.getInnerTilePolygonAt( i==0 ? 9 : i-1 ).vertices[0].clone() );
-	
-
 	this.outerTilePolygons.push( outerTile );
     }
-    
 };
 
 
 // This is totally shitty. Why object inheritance when I still
 // have to inherit object methods manually??!
 GirihDecagon.prototype.computeBounds         = GirihTile.prototype.computeBounds;
-GirihDecagon.prototype._addVertex            = GirihTile.prototype._addVertex;
-GirihDecagon.prototype._translateVertex      = GirihTile.prototype._translateVertex;
+GirihDecagon.prototype.addVertex             = GirihTile.prototype.addVertex;
+GirihDecagon.prototype.translateVertex       = GirihTile.prototype.translateVertex;
 GirihDecagon.prototype._polygonToSVG         = GirihTile.prototype._polygonToSVG;
 GirihDecagon.prototype.getInnerTilePolygonAt = GirihTile.prototype.getInnerTilePolygonAt;
 GirihDecagon.prototype.getOuterTilePolygonAt = GirihTile.prototype.getOuterTilePolygonAt;
@@ -120,7 +134,7 @@ GirihDecagon.prototype.getTranslatedVertex   = GirihTile.prototype.getTranslated
 GirihDecagon.prototype.containsPoint         = GirihTile.prototype.containsPoint;
 GirihDecagon.prototype.locateEdgeAtPoint     = GirihTile.prototype.locateEdgeAtPoint;
 GirihDecagon.prototype.locateAdjacentEdge    = GirihTile.prototype.locateAdjacentEdge;
-GirihDecagon.prototype.getVertexAt           = GirihTile.prototype.getVertexAt;
+GirihDecagon.prototype.getVertexAt          = Polygon.prototype.getVertexAt; // GirihTile.prototype.getVertexAt;
 GirihDecagon.prototype.toSVG                 = GirihTile.prototype.toSVG;
 
 GirihDecagon.prototype.constructor           = GirihDecagon;