+++ docs_jekyll/changelog.md
@@ -1,11 +1,15 @@
 ---
 layout: page
+title: Changelog
 ---
 
 
 # Changelog
-* 2020-10-27
-  * Added the draw.svgPath(...) function.
+* 2020-10-31
+  * Including utils in the docs now.
+* 2020-10-30
+  * Added the static Bounds.computeFromVertices(Vertex[]) function.
+  * Added the `Polygon.addVert` function.
 * 2020-10-23
   * Added the fillArray and fillMatrix helper function (like in lodash).
   * Added the `Matrix`, `Interval` and `IndexPair` interfaces to the typescript definitions.